api/requirements.txt
uvicorn[standard]==0.30.1
gunicorn==22.0.0
fastapi==0.111.1
pydantic==2.8.2
jsonschema==4.6.0
loguru==0.7.2
requests==2.32.3
python-keycloak==4.2.2
# asyncio==3.4.3
pytz==2024.1
psycopg2-binary==2.9.10